fix copy to clipboard and change int to BigInteger expression type

+++ app/src/main/java/com/matheusvalbert/programmercalculator/core/usecase/expression/EvaluatorUseCase.java
@@ -3,6 +3,8 @@ package com.matheusvalbert.programmercalculator.core.usecase.expression;
 import com.matheusvalbert.programmercalculator.core.usecase.UseCase;
 import com.matheusvalbert.programmercalculator.core.util.Base;
 
+import java.math.BigInteger;
+
 public interface EvaluatorUseCase extends UseCase {
-    int invoke(String expression, @Base int base);
+    BigInteger invoke(String expression, @Base int base);
 }
