+++ app/src/main/java/com/matheusvalbert/programmercalculator/core/usecase/expression/EvaluatorUseCaseImpl.java
@@ -4,6 +4,7 @@ import com.matheusvalbert.programmercalculator.core.domain.Evaluator;
 import com.matheusvalbert.programmercalculator.core.domain.Tokenizer;
 import com.matheusvalbert.programmercalculator.core.util.Base;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public class EvaluatorUseCaseImpl implements EvaluatorUseCase {
@@ -17,7 +18,7 @@ public class EvaluatorUseCaseImpl implements EvaluatorUseCase {
     }
 
     @Override
-    public int invoke(String expression, @Base int base) {
+    public BigInteger invoke(String expression, @Base int base) {
         List<String> tokens = mTokenizer.tokenize(expression, base);
         return mEvaluator.evaluate(tokens);
     }
